Exit when downstream pipe closes
For awhile I have been getting error messages on commands like

$ smof sort seq.faa | smof head `smof head`

They do not affect the final results or impact performance, so I have
been ignoring them. But they are annoying and may, with good reason,
dismay users.

These errors occur when data is sent to a pipe that has been closed. In
the example above, `smof head` prints and closes after receiving the
first sequence, so `smof sort` fails when it attempts to send a second.
I've fixed this problem by exiting gracefully (error code of 0) when
the BrokenPipeError exeption is caught.