Remove -H --with-filename command
This option was added in the previous commit. Its only function was to
add the filename as a tag to the header. This could be useful if you
wanted to separate input files back into their original groups after
processing, but as it was implemented it did not play nicely with other
options (like -o and --gff). Maybe in a future release I can bring it
back. Perhaps adding a function to split that allows splitting into
files determined by regex matches to the header. But that is fancy stuff
for later.