Store results of guess_type
Previously, when many bounds were passed to Subseq, the type (e.g. dna
or prot) was guessed for each bound. This lead to incredibly slow
processivity when, for example, large GFF files were extracted from
genome sequences. Now guess_type is only run once and the result is
stored in the FSeq objects.