Allow empty sequences
Empty sequences, such as:

```
> foo
> bar
1 2 3
```

are generally a BAD THING. But they do happen in the wild. Further, they
could reasonably be returned from programs that, for example, contain
matches of a query against a database. The empty lists might mean no
result was found for the given header. So they should be supported.