Add tests and fix start handling
I also factored the translation handling into a dedicated function. I
should actually consider doing this all through smof, since the pure
functions are easier to test and can be reused if smof is imported as a
module.

The start handling is now performed only when all frames is used as
well. Requiring a start while also searching only the first frame is
probably not something one should do -- if there are additional
nucleotides before the start codon, why would they just happen to occur
in a multiple of 3?