Add --replace-header option to md5sum
`smof md5sum --replace-header` will replace the headers of each entry in
a file with the md5 chksum of the sequence. This is useful in a system
that needs a unique ID for every sequence, for example where the md5
chksum is used as an identifier linking to other sequence data.