Drop testing of python 3.4, 3.3 and 3.2
Now I test only on 3.5 and 3.6. Travis was complaining that it could
find 3.3 and one of the tests failed on 3.4. The test failed due to a
difference in the order entries were printed in by `smof uniq`. This is
not a serious problem, since `smof uniq` does not claim to preserve
input order.